Refactor agent details page and remove unused components
